\name{MGGM-package}
\alias{MGGM-package}
\alias{MGGM}
\docType{package}
\title{
\packageTitle{MGGM}
}

\description{
\packageDescription{MGGM}
}

\details{
\packageDESCRIPTION{MGGM}
\packageIndices{MGGM}
}

\author{
\packageAuthor{MGGM}
Maintainer: \packageMaintainer{MGGM}
}

\references{
Yunzhang Zhu, Xiaotong Shen, Wei Pan. Structural pursuit over multiple undirected graphs
}